Dashboard Config (#176)
* Update dashboard to use grouping

* Dashboard config

* Update frontend admin list code

* Adding indicators & context layer form of dashboard config

* Dashboard Config : Filters

* Dashboard config : Widgets

* Fix flake

* Dashboard Config: Add preview

* Dashboard Config: Remove edit

* Dashboard Config: Save Dashboard

* Fix flake

+++ django_project/frontend/src/pages/Admin/Components/List/index.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 import $ from "jquery";
 import Tooltip from '@mui/material/Tooltip';
 import SearchIcon from '@mui/icons-material/Search';
@@ -7,9 +7,7 @@ import { GridActionsCellItem } from "@mui/x-data-grid";
 import EditIcon from "@mui/icons-material/Edit";
 import DeleteIcon from "@mui/icons-material/Delete";
 
-import Admin from '../../index';
 import { AdminTable } from '../Table';
-import { AddButton } from '../../../../components/Elements/Button'
 import { IconTextField } from '../../../../components/Elements/Input'
 import { fetchingData } from "../../../../Requests";
 
@@ -28,44 +26,53 @@ import './style.scss';
 export function COLUMNS_ACTION(params, redirectUrl, editUrl = null, detailUrl = null) {
   editUrl = editUrl ? editUrl : urls.api.edit;
   detailUrl = detailUrl ? detailUrl : urls.api.detail;
-  return [
-    <GridActionsCellItem
-      icon={
-        <Tooltip title={`Edit ${params.row.name}`}>
-          <a
-            href={editUrl.replace('/0', `/${params.id}`)}>
-            <EditIcon/>
-          </a>
-        </Tooltip>
-      }
-      label="Edit"
-    />,
-    <GridActionsCellItem
-      className='AdminTableDelete'
-      icon={
-        <Tooltip title={`Delete ${params.row.name}`}>
-          <DeleteIcon/>
-        </Tooltip>
-      }
-      label="Delete"
-      onClick={
-        () => {
-          const api = detailUrl.replace('/0', `/${params.id}`);
-          if (confirm(`Are you sure you want to delete : ${params.row.name}?`)) {
-            $.ajax({
-              url: api,
-              method: 'DELETE',
-              success: function () {
-                window.location = redirectUrl;
-              },
-              beforeSend: beforeAjaxSend
-            });
-            return false;
+
+  const actions = []
+  if (editUrl) {
+    actions.push(
+      <GridActionsCellItem
+        icon={
+          <Tooltip title={`Edit ${params.row.name}`}>
+            <a
+              href={editUrl.replace('/0', `/${params.id}`)}>
+              <EditIcon/>
+            </a>
+          </Tooltip>
+        }
+        label="Edit"
+      />
+    )
+  }
+  if (detailUrl) {
+    actions.push(
+      <GridActionsCellItem
+        className='AdminTableDelete'
+        icon={
+          <Tooltip title={`Delete ${params.row.name}`}>
+            <DeleteIcon/>
+          </Tooltip>
+        }
+        label="Delete"
+        onClick={
+          () => {
+            const api = detailUrl.replace('/0', `/${params.id}`);
+            if (confirm(`Are you sure you want to delete : ${params.row.name}?`)) {
+              $.ajax({
+                url: api,
+                method: 'DELETE',
+                success: function () {
+                  window.location = redirectUrl;
+                },
+                beforeSend: beforeAjaxSend
+              });
+              return false;
+            }
           }
         }
-      }
-    />,
-  ]
+      />
+    )
+  }
+  return actions
 }
 
 /**
@@ -84,7 +91,7 @@ export function COLUMNS(pageName, redirectUrl, editUrl = null, detailUrl = null)
     { field: 'id', headerName: 'id', hide: true, width: 30 },
     { field: 'name', headerName: pageName + ' Name', flex: 1 },
     { field: 'description', headerName: 'Description', flex: 1 },
-    { field: 'group', headerName: 'Category', flex: 1 },
+    { field: 'category', headerName: 'Category', flex: 1 },
     {
       field: 'actions',
       type: 'actions',
@@ -101,22 +108,23 @@ export function COLUMNS(pageName, redirectUrl, editUrl = null, detailUrl = null)
  * @param {list} columns Columns setup.
  * @param {String} pageName Page Name.
  * @param {String} listUrl Url for list row.
- * @param {React.Component} children React component to be rendered
+ * @param {list} initData Init Data.
  */
-export default function AdminList(
+export default function List(
   {
-    columns, pageName,
-    listUrl, rightHeader = null
+    columns, pageName, listUrl, initData
   }
 ) {
-  const [data, setData] = useState(null);
+  const [data, setData] = useState(initData);
   const [search, setSearch] = useState(null);
 
   /** Fetch list of data */
   const fetchData = (url) => {
-    fetchingData(url, {}, {}, (data) => {
-      setData(data)
-    })
+    if (!data) {
+      fetchingData(url, {}, {}, (data) => {
+        setData(data)
+      })
+    }
   }
 
   // Show modal when url changed
@@ -134,27 +142,14 @@ export default function AdminList(
   if (search) {
     rows = rows.filter(row => {
       return row.name.toLowerCase().includes(search)
-        || row.group.toLowerCase().includes(search)
+        || row.category.toLowerCase().includes(search)
         || row.description.toLowerCase().includes(search)
     })
   }
 
   /** Render **/
   return (
-    <Admin
-      className='Indicator'
-      pageName={pageName}
-      rightHeader={
-        rightHeader ? rightHeader : (
-          <a href={urls.api.create}>
-            <AddButton
-              variant="secondary"
-              text={"Add New " + pageName}
-            />
-          </a>
-        )
-      }>
-
+    <Fragment>
       <div className='AdminBaseInput Indicator-Search'>
         <IconTextField
           placeholder={"Search " + pageName}
@@ -167,6 +162,6 @@ export default function AdminList(
         <AdminTable
           rows={rows} columns={columns}/>
       </div>
-    </Admin>
+    </Fragment>
   );
 }
